Fix test-flock.cc

+++ test/test-flock.cc
@@ -5,45 +5,43 @@
 #include "string.h"
 #include "gtest/gtest.h"
 #include <sys/file.h>
+#include <sys/wait.h>
 
 namespace {
 
 TEST(FlockTest, BasicFileLocking) {
   char template_name[] = "/tmp/test_file_lock_XXXXXX";
   int fd = mkstemp(template_name);
-  ASSERT_NE(fd, 0) << "Failed to create a unique temporary file";
+  ASSERT_NE(fd, -1) << "Failed to create a unique temporary file";
 
   int ret = flock(fd, LOCK_EX);
   EXPECT_EQ(ret, 0) << "Failed to acquire exclusive lock";
 
-  int sharedRet = flock(fd, LOCK_SH | LOCK_NB);
-  EXPECT_NE(sharedRet, 0) << "Should not be able to acquire shared lock when exclusive is held";
+  pid_t pid = fork();
 
-  ret = flock(fd, LOCK_UN);
-  EXPECT_EQ(ret, 0) << "Failed to release lock";
+  if (pid == 0) { // Child process, should fail to lock fd
+    int childRet = flock(fd, LOCK_EX | LOCK_NB);
+    EXPECT_EQ(childRet, -1) << "Child should fail to acquire exclusive lock";
+      
+    if (childRet == 0) {
+      exit(EXIT_FAILURE);
+    }
+      
+    exit(EXIT_SUCCESS);
+  } else if (pid > 0) { // Parent process
+      int status;
+      pid_t childPid = wait(&status);
+        
+      EXPECT_EQ(WEXITSTATUS(status), EXIT_SUCCESS) << "Child process did not exit with success status";
+      
+      ret = flock(fd, LOCK_UN);
+      EXPECT_EQ(ret, 0) << "Failed to release lock";
 
-  close(fd);
-
-  unlink(filename);
-}
-
-TEST(FlockTest, NonBlockingLock) {
-  char template_name[] = "/tmp/test_file_lock_nonblocking_XXXXXX";
-  int fd = mkstemp(template_name);
-  ASSERT_NE(fd, -1) << "Failed to create a unique temporary file";
-
-  int ret = flock(fd, LOCK_EX | LOCK_NB);
-  EXPECT_EQ(ret, 0) << "Failed to acquire non-blocking exclusive lock";
-
-  int ret2 = flock(fd, LOCK_EX | LOCK_NB);
-  EXPECT_NE(ret2, 0) << "Should not be able to acquire non-blocking exclusive lock again";
-
-  ret = flock(fd, LOCK_UN);
-  EXPECT_EQ(ret, 0) << "Failed to release lock";
-
-  close(fd);
-
-  unlink(template_name);
+      close(fd);
+      unlink(template_name);
+  } else { // Fork failed
+      FAIL() << "Failed to fork a child process";
+  }
 }
 
 } // namespace
